Add a LICENSE file
The chosen MIT license requires including the copyright and permission
notices (license text) “in all copies or substantial portions of the
Software,” so the license text needs to be present in the published
crates.

The LICENSE file is based on the one in https://github.com/pdf-rs/pdf/,
with the copyright date adjusted from 2020 to 2021 since that is the
date of the initial commit, 9eb83d64d7c48e10b2fe864906f0b9e67f83c0ef.